Function krew_wasm_plugin_sdk::wasi_outbound_http_helper_k8s::make_request

pub fn make_request(
    k8s_req: Request<Vec<u8>>, 
    connection_config: &ConnectionConfig, 
    req_cfg_id: &str
) -> Result<Response>

Perform HTTP request using the given connection configuration object and the given request configuration

The http:Request objects created by the k8s-openapi crate do not have a target host specified. This function takes care of that by looking at the Kubernetes settings provided by the ConnectionConfig object.